linux testing

+++ library/core/source/allio/impl/linux/pidfd.hpp
@@ -15,7 +15,7 @@ namespace allio::linux {
 
 inline vsm::result<detail::unique_handle> pidfd_open(pid_t const pid, unsigned const flags)
 {
-	int const fd = syscall(SYS_pidfd_open, pid, flags);
+	int const fd = static_cast<int>(syscall(SYS_pidfd_open, pid, flags));
 
 	if (fd == -1)
 	{
@@ -25,9 +25,13 @@ inline vsm::result<detail::unique_handle> pidfd_open(pid_t const pid, unsigned c
 	return vsm_lazy(detail::unique_handle(fd));
 }
 
-inline int pidfd_send_signal(int const fd, int const sig, siginfo_t* const info, unsigned const flags)
+inline int pidfd_send_signal(
+	int const fd,
+	int const sig,
+	siginfo_t* const info,
+	unsigned const flags)
 {
-	return syscall(SYS_pidfd_send_signal, fd, sig, info, flags);
+	return static_cast<int>(syscall(SYS_pidfd_send_signal, fd, sig, info, flags));
 }
 
 } // namespace allio::linux
